pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["Elizaveta Latypova <elizaveta1902@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
python-multipart = "^0.0.5"
SQLAlchemy = "^1.4.25"
dynaconf = "^3.1.7"
aiofiles = "^0.7.0"
passlib = "^1.7.4"
alembic = "^1.7.3"
itsdangerous = "^2.0.1"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
